[HOUSEKEEPING] Removed commented out metods (see desc)
Prev. commit was a mistake I will make a long post about it and it's implications for the community later but for now be aware that this event has changed me as a person and from now on I will do everything I can to prevent similar events from happening.

I hope we can all move on from this despite the damage that has been done to the community